Add an documentation to the SimpleNesting
This commit adds a documentation to SimpleNesting as well as modifies
its deploy script to simply deploy it. User journey script has been
added to replace the showcase of SimpleNesting usage.

Custom script for running the SimpleNesting user journey has been added.

Nesting example smart contracts have been moved into their own folder,
so that the documentation of each architecture can be separate from the
main README. This should increase readability and help us avoid a
monolithic README.
