+++ contracts/README.md
@@ -3,3 +3,5 @@
 This section contains the examples of using RMRK legos to build your own smart contracts. Every example uses the `@rmrk-team/evm-contracts` dependency in order to implement the examples. Adding it allows you to easily include them in yout own smart contracts.
 
 The examples covered are:
+
+1. [Nesting](./Nesting/README.md) examines the Nesting RMRK lego.
