+++ package.json
@@ -8,7 +8,8 @@
     "deploy-multiresource": "hardhat run scripts/deployMultiresource.ts",
     "deploy-nesting": "hardhat run scripts/deployNesting.ts",
     "deploy-nesting-multiresource": "hardhat run scripts/deployNestingMultiresource.ts",
-    "deploy-split-equippable": "hardhat run scripts/deploySplitEquippable.ts"
+    "deploy-split-equippable": "hardhat run scripts/deploySplitEquippable.ts",
+    "user-journey-nesting": "hardhat run scripts/nestingUserJourney.ts"
   },
   "keywords": [],
   "author": "@rmrk-team",
